+++ ClickstreamTests/EventProcessorTests/CourierEventProcessorTest.swift
@@ -15,6 +15,7 @@ class CourierEventProcessorTest: XCTestCase {
     private var mockQueue: SerialQueue!
     private var mockClassifier: MockEventClassifier!
     private var mockWarehouser: CourierEventWarehouser!
+    private var mockSampler: EventSampler!
     private var courierEventProcessor: CourierEventProcessor!
     private var courierBatchEventProcessor: CourierEventBatchProcessor!
     private var courierBatchCreator: CourierEventBatchCreator!
@@ -57,8 +58,8 @@ class CourierEventProcessorTest: XCTestCase {
         courierEventProcessor = CourierEventProcessor(
             performOnQueue: mockQueue,
             classifier: mockClassifier,
-            eventWarehouser: mockWarehouser,
-            networkOptions: networkOptions
+            eventWarehouser: mockWarehouser, sampler: mockSampler,
+            networkOptions: networkOptions, eventExpiryManager: FallbackEventExpirationManager()
         )
 
         testEvent = ClickstreamEvent(
@@ -86,8 +87,8 @@ class CourierEventProcessorTest: XCTestCase {
         courierEventProcessor = CourierEventProcessor(
             performOnQueue: mockQueue,
             classifier: mockClassifier,
-            eventWarehouser: mockWarehouser,
-            networkOptions: networkOptions
+            eventWarehouser: mockWarehouser, sampler: mockSampler,
+            networkOptions: networkOptions, eventExpiryManager: FallbackEventExpirationManager()
         )
         
         XCTAssertNotNil(courierEventProcessor)
@@ -97,8 +98,8 @@ class CourierEventProcessorTest: XCTestCase {
         courierEventProcessor = CourierEventProcessor(
             performOnQueue: mockQueue,
             classifier: mockClassifier,
-            eventWarehouser: mockWarehouser,
-            networkOptions: networkOptions
+            eventWarehouser: mockWarehouser, sampler: mockSampler,
+            networkOptions: networkOptions, eventExpiryManager: FallbackEventExpirationManager()
         )
         
         XCTAssertNotNil(courierEventProcessor)
@@ -113,8 +114,8 @@ class CourierEventProcessorTest: XCTestCase {
         courierEventProcessor = CourierEventProcessor(
             performOnQueue: mockQueue,
             classifier: mockClassifier,
-            eventWarehouser: mockWarehouser,
-            networkOptions: networkOptions
+            eventWarehouser: mockWarehouser, sampler: mockSampler,
+            networkOptions: networkOptions, eventExpiryManager: FallbackEventExpirationManager()
         )
 
         XCTAssertFalse(courierEventProcessor.shouldTrackEvent(event: testEvent))